requirements.txt
flask >= 0.11.1
psycopg2 >= 2.6.2
uwsgi >= 2.0.13.1
pylibmc >= 1.5.1
dnspyhton >= 1.14.0